Add truncatedAlerts field to webhook message

+++ config/notifiers.go
@@ -412,7 +412,7 @@ type WebhookConfig struct {
 	// MaxAlerts is the maximum number of alerts to be sent per webhook message.
 	// Alerts exceeding this threshold will be truncated. Setting this to 0
 	// allows an unlimited number of alerts.
-	MaxAlerts uint32 `yaml:"max_alerts" json:"max_alerts"`
+	MaxAlerts uint64 `yaml:"max_alerts" json:"max_alerts"`
 }
 
 // UnmarshalYAML implements the yaml.Unmarshaler interface.
